+++ inst/doc/deployment.R
@@ -42,6 +42,8 @@
 #  DW.yaml       = system.file(package="formods",  "templates",  "DW.yaml")
 #  FG.yaml       = system.file(package="formods",  "templates",  "FG.yaml")
 #  NCA.yaml      = system.file(package="ruminate", "templates",  "NCA.yaml")
+#  MB.yaml       = system.file(package="ruminate", "templates",  "MB.yaml")
+#  CTS.yaml      = system.file(package="ruminate", "templates",  "CTS.yaml")
 
 ## ----eval=FALSE---------------------------------------------------------------
 #  file.copy(from = system.file(package="ruminate", "templates",  "NCA.yaml"),
@@ -49,4 +51,22 @@
 
 ## ----eval=FALSE---------------------------------------------------------------
 #  NCA.yaml      = "myNCA.yaml"
+
+## ----eval=FALSE---------------------------------------------------------------
+#  file.copy(from = system.file(package="formods", "templates",  "formods.yaml"),
+#            to   = "myformods.yaml")
+
+## ----eval=FALSE---------------------------------------------------------------
+#  FM:
+#    include:
+#      files:
+#      - file:
+#          source: 'file.path(".", "templates", "report.docx")'
+#          dest:   'file.path("config","report.docx")'
+#      - file:
+#          source: 'file.path(".", "templates", "report.pptx")'
+#          dest:   'file.path("config","report.pptx")'
+#      - file:
+#          source: 'file.path(".", "templates", "report.yaml")'
+#          dest:   'file.path("config","report.yaml")'