+++ NEWS.md
@@ -1,6 +1,13 @@
 # ruminate 0.2.2 (development version)
 
+- Added default naming of NCA analysis from datasets
+- Added searching in selection of NCA parameters
+- Fixed sparse sampling in NCA
 - Added CTS (Clinical trial simulator) module (in development)
+
+# ruminate 0.2.2 
+
+- Fixed error in default ruminate app causing it to crash
 - Separated ruminate.R sample app into two separate (the default one and one for developent.)
 
 # ruminate 0.2.1